Turney Algorithm for Sentiment Classification
This project implements the Turney Algorithm for sentiment
classification on the IMDB dataset.
Turney Algorithm is an unsupervised sentiment classification based on
phrasal lexicon.
Phrases are selected based on grammatical rules associated with their
POS tags, ( based on the paper (link:
http://acl-arc.comp.nus.edu.sg/archives/acl-arc-090501d4/data/pdf/anthology-PDF/P/P02/P02-1053.pdf).

The semantic orientation of each phrase is calculated to the number of
words of "great" and "poor" near it in the dataset.
The polarity of the review is the sum total of the semantic orientation
of the phrases in the document. The document is classified positive if
the polarity is positive and negative if the polarity is negative